Add app-shell container file
:Release Notes:
Add a container file to register app-shell as a container type

:Detailed Notes:
An app-shell is one of browsers in chromium and its extension is enact
browser in OSE. For ACG, app-shell should be a container like wam. So,
this patch added com.webos.app-shell.container.json file.

:Testing Performed:
Tested with verification build image.

:QA Notes:

:Issues Addressed:
[PLAT-67708] [OSE][ACG][App-shell] Implement Client side of Luna Service
as an application container

Change-Id: I3c9d17d8faffc504bfbebd5df3afe0e198000fb8